requirements.txt
iracingdataapi>=1.2.3
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
strawberry-graphql[fastapi]>=0.215.0
sqlalchemy>=2.0.0
uvicorn[standard]>=0.24.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
python-dotenv>=1.0.0
litecli>=1.10.0